Move ReflectionExtensions to /Common
